feat: Complete assignment steps 1-5
1. Declare `ng-app` either on the `html` or the `body` element.
Name your app `NarrowItDownApp`.

2. Create `app.js` in your project and declare an Angular module
to match your `ng-app` declaration.

3. Declare and create a NarrowItDownController (with controller as syntax)
that will wrap your search textbox and button as well as the list of found items.

4. Declare and create MenuSearchService. The service should have the following
method: getMatchedMenuItems(searchTerm). That method will be responsible for
reaching out to the server (using the $http service) to retrieve the list
of all the menu items. Once it gets all the menu items, it should loop
through them to pick out the ones whose description matches the searchTerm.
Once a list of found items is compiled, it should return that list (wrapped
in a promise). Remember that the `then` function itself returns a promise.

5. The NarrowItDownController should be injected with the MenuSearchService.
The controller should call the getMatchedMenuItems method when appropriate
and store the result in a property called `found` attached to the controller
instance.