feat: Complete assignment step 6
6. Declare and create `foundItems` directive. The list should be displayed
using this `directive` which takes the `found` array of items specified on it
as an attribute (think one-way binding with '<'). To implement
the functionality of the "Don't want this one!" button, the directive
should also provide an on-remove attribute that will use
function reference binding to invoke the parent controller removal an item
from the `found` array based on an index into the `found` array. The index
should be passed in from the directive to the controller.
In the NarrowItDownController, simply remove that item from the `found` array.